Replace abstract factory with factory method.
The abstract ReposoitoryFactory was overly complicating the situation
that is implemented quite nicely through a simpler Factory Method.  The
implmenetation of SimpleDataNamedConnectionDatabaseProvider is
initialised with the name of the named connection to open.  It could be
made a bit smarter to scope the connections on a per-request basis so
the connection can be reused between repositories within the same web
request.