Add tests for unsupported Membership return codes.
UserRejected, [Duplicate|Invalid]ProviderUserKey,
Invalid[Question|Answer] and ProviderError status codes are unsupported
by the AspMembershipService and are represented as a catch-all
UnexpectedFailure. This is because we use GUIDs as the Provider User Key
(as expected by the default SQL Membership Provider), which are
guaranteed to be globally unique; the membership provider is configured
to disable all features requiring secret questions and answers, and
UserRejected and ProviderError do not provide much useful information on
their own.